add postFeed.lib.js to seperate methods and logics
