+++ TO_KEEP/postFeed.html
@@ -4,6 +4,6 @@
 </head>
 <body>
     <canvas id="canvas"></canvas>
-    <script src="postFeed.js"></script>
+    <script type="module" src="postFeed.js"></script>
 </body>
 </html>
