+++ TO_KEEP/postFeed.js
@@ -1,3 +1,4 @@
+import {Vec2} from "./postFeed.lib.js";
 const canvas = document.getElementById('canvas');
 const ctx = canvas.getContext('2d');
 const ROTATION_FACTOR = 1/75;
@@ -14,23 +15,7 @@ const ROTATION_INITIAL_MOMENTUM_VERTICAL = 0;
 canvas.width = window.innerWidth;
 canvas.height = window.innerHeight;
 
-class Vec2 {
-    constructor(x,y){
-        this.x = x;
-        this.y = y;
-    }
-    static Magnitude(v){
-        const px = Math.pow(v.x,2);
-        const py = Math.pow(v.y,2);
-        const magnitude = Math.sqrt(px+py);
-        return magnitude;
-    }
-    static Delta(vMinuend,vSubtrahend){
-        const dx = vMinuend.x - vSubtrahend.x;
-        const dy = vMinuend.y - vSubtrahend.y;
-        return new Vec2(dx,dy);
-    }
-}
+
 
 class Vec3 {
     constructor(x,y,z){
